Refactor: 네트워크 프로토콜화를 통한 SOLID 원칙 준수
- SRP, ISP: 디코딩을 하는 객체와 디코딩을 하지 않는 객체를 분리를 통해 단일 책임을 갖는 인터페이스로 분리
- OCP, LSP, DIP: 프로토콜화를 하여 프로토콜을 따르는 객체를 추가로 생성하여 OCP를 준수, 외부에서 객체를 주입할 때 프로토콜을 따르는 객체라면 그 자리를 대신할 수 있게 하여 LSP와 DIP를 준수할 수 있다.
